[GCP] Move antispam code into own package
