+++ storage/gcp/gcp.go
@@ -38,7 +38,6 @@ import (
 	"net/http"
 	"os"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"cloud.google.com/go/spanner"
@@ -1021,310 +1020,6 @@ func (s *gcsStorage) lastModified(ctx context.Context, obj string) (time.Time, e
 		return time.Time{}, fmt.Errorf("failed to create reader for object %q in bucket %q: %w", obj, s.bucket, err)
 	}
 	return r.Attrs.LastModified, r.Close()
-}
-
-// NewAntispam returns an antispam driver which uses Spanner to maintain a mapping of
-// previously seen entries and their assigned indices.
-//
-// Note that the storage for this mapping is entirely separate and unconnected to the storage used for
-// maintaining the Merkle tree.
-//
-// This functionality is experimental!
-func NewAntispam(ctx context.Context, spannerDB string) (*AntispamStorage, error) {
-	if err := createAndPrepareTables(
-		ctx, spannerDB,
-		[]string{
-			"CREATE TABLE IF NOT EXISTS FollowCoord (id INT64 NOT NULL, nextIdx INT64 NOT NULL) PRIMARY KEY (id)",
-			"CREATE TABLE IF NOT EXISTS IDSeq (id INT64 NOT NULL, h BYTES(32) NOT NULL, idx INT64 NOT NULL) PRIMARY KEY (id, h)",
-		},
-		[][]*spanner.Mutation{
-			{spanner.Insert("FollowCoord", []string{"id", "nextIdx"}, []interface{}{0, 0})},
-		},
-	); err != nil {
-		return nil, fmt.Errorf("failed to create tables: %v", err)
-	}
-
-	db, err := spanner.NewClient(ctx, spannerDB)
-	if err != nil {
-		return nil, fmt.Errorf("failed to connect to Spanner: %v", err)
-	}
-
-	r := &AntispamStorage{
-		dbPool: db,
-	}
-
-	go func(ctx context.Context) {
-		t := time.NewTicker(time.Second)
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case <-t.C:
-				klog.V(1).Infof("ANTISPAM: # Writes %d, # Lookups %d, # DB hits %v", r.numWrites.Load(), r.numLookups.Load(), r.numHits.Load())
-			}
-		}
-	}(ctx)
-
-	return r, nil
-}
-
-type AntispamStorage struct {
-	dbPool *spanner.Client
-
-	// pushBack is used to prevent the follower from getting too far underwater.
-	// Populate dynamically will set this to true/false based on how far behind the follower is from the
-	// currently integrated tree size.
-	// When pushBack is true, the decorator will start returning ErrPushback to all calls.
-	pushBack atomic.Bool
-
-	numLookups atomic.Uint64
-	numWrites  atomic.Uint64
-	numHits    atomic.Uint64
-}
-
-// index returns the index (if any) previously associated with the provided hash
-func (d *AntispamStorage) index(ctx context.Context, h []byte) (*uint64, error) {
-	d.numLookups.Add(1)
-	var idx int64
-	if row, err := d.dbPool.Single().ReadRow(ctx, "IDSeq", spanner.Key{0, h}, []string{"idx"}); err != nil {
-		if c := spanner.ErrCode(err); c == codes.NotFound {
-			return nil, nil
-		}
-		return nil, err
-	} else {
-		if err := row.Column(0, &idx); err != nil {
-			return nil, fmt.Errorf("failed to read antispam index: %v", err)
-		}
-		idx := uint64(idx)
-		d.numHits.Add(1)
-		return &idx, nil
-	}
-}
-
-// Decorator returns a function which will wrap an underlying Add delegate with
-// code to dedup against the stored data.
-func (d *AntispamStorage) Decorator() func(f tessera.AddFn) tessera.AddFn {
-	return func(delegate tessera.AddFn) tessera.AddFn {
-		return func(ctx context.Context, e *tessera.Entry) tessera.IndexFuture {
-			if d.pushBack.Load() {
-				// The follower is too far behind the currently integrated tree, so we're going to push back against
-				// the incoming requests.
-				// This should have two effects:
-				//   1. The tree will cease growing, giving the follower a chance to catch up, and
-				//   2. We'll stop doing lookups for each submission, freeing up Spanner CPU to focus on catching up.
-				//
-				// We may decide in the future that serving duplicate reads is more important than catching up as quickly
-				// as possible, in which case we'd move this check down below the call to index.
-				return func() (uint64, error) { return 0, tessera.ErrPushback }
-			}
-			idx, err := d.index(ctx, e.Identity())
-			if err != nil {
-				return func() (uint64, error) { return 0, err }
-			}
-			if idx != nil {
-				return func() (uint64, error) { return *idx, nil }
-			}
-
-			return delegate(ctx, e)
-		}
-	}
-}
-
-// entryStreamReader converts a stream of {RangeInfo, EntryBundle} into a stream of individually processed entries.
-//
-// TODO(al): Factor this out for re-use elsewhere when it's ready.
-type entryStreamReader[T any] struct {
-	bundleFn func([]byte) ([]T, error)
-	next     func() (layout.RangeInfo, []byte, error)
-
-	curData []T
-	curRI   layout.RangeInfo
-	i       uint64
-}
-
-// newEntryStreamReader creates a new stream reader which uses the provided bundleFn to process bundles into processed entries of type T
-//
-// Different bundleFn implementations can be provided to return raw entry bytes, parsed entry structs, or derivations of entries (e.g. hashes) as needed.
-func newEntryStreamReader[T any](next func() (layout.RangeInfo, []byte, error), bundleFn func([]byte) ([]T, error)) *entryStreamReader[T] {
-	return &entryStreamReader[T]{
-		bundleFn: bundleFn,
-		next:     next,
-		i:        0,
-	}
-}
-
-// Next processes and returns the next available entry in the stream along with its index in the log.
-func (e *entryStreamReader[T]) Next() (uint64, T, error) {
-	var t T
-	if len(e.curData) == 0 {
-		var err error
-		var b []byte
-		e.curRI, b, err = e.next()
-		if err != nil {
-			return 0, t, fmt.Errorf("next: %v", err)
-		}
-		e.curData, err = e.bundleFn(b)
-		if err != nil {
-			return 0, t, fmt.Errorf("bundleFn(bundleEntry @%d): %v", e.curRI.Index, err)
-
-		}
-		if e.curRI.First > 0 {
-			e.curData = e.curData[e.curRI.First:]
-		}
-		if len(e.curData) > int(e.curRI.N) {
-			e.curData = e.curData[:e.curRI.N]
-		}
-		e.i = 0
-	}
-	t, e.curData = e.curData[0], e.curData[1:]
-	rIdx := e.curRI.Index*layout.EntryBundleWidth + uint64(e.curRI.First) + e.i
-	e.i++
-	return rIdx, t, nil
-}
-
-// Populate uses entry data from the log to populate the antispam storage.
-//
-// TODO(al):  add details
-func (d *AntispamStorage) Populate(ctx context.Context, lr tessera.LogReader, bundleFn func([]byte) ([][]byte, error)) {
-	errOutOfSync := errors.New("out-of-sync")
-
-	t := time.NewTicker(time.Second)
-	var (
-		entryReader *entryStreamReader[[]byte]
-		stop        func()
-
-		curEntries [][]byte
-		curIndex   uint64
-	)
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-t.C:
-		}
-		size, err := lr.IntegratedSize(ctx)
-		if err != nil {
-			klog.Errorf("Populate: IntegratedSize(): %v", err)
-			continue
-		}
-
-		// Busy loop while there's work to be done
-		for workDone := true; workDone; {
-			_, err = d.dbPool.ReadWriteTransaction(ctx, func(ctx context.Context, txn *spanner.ReadWriteTransaction) error {
-				// Figure out the last entry we used to populate our antispam storage.
-				row, err := txn.ReadRowWithOptions(ctx, "FollowCoord", spanner.Key{0}, []string{"nextIdx"}, &spanner.ReadOptions{LockHint: spannerpb.ReadRequest_LOCK_HINT_EXCLUSIVE})
-				if err != nil {
-					return err
-				}
-
-				var f int64 // Spanner doesn't support uint64
-				if err := row.Columns(&f); err != nil {
-					return fmt.Errorf("failed to read follow coordination info: %v", err)
-				}
-				followFrom := uint64(f)
-				if followFrom >= size {
-					// Our view of the log is out of date, exit the busy loop and refresh it.
-					workDone = false
-					return nil
-				}
-
-				// TODO(al): Maybe make these configurable.
-				const batchSize = 40
-				const pushBackThreshold = batchSize * 16
-				d.pushBack.Store(size-followFrom > pushBackThreshold)
-
-				// If this is the first time around the loop we need to start the stream of entries now that we know where we want to
-				// start reading from:
-				if entryReader == nil {
-					next, st := lr.StreamEntries(ctx, followFrom)
-					stop = st
-					entryReader = newEntryStreamReader(next, bundleFn)
-				}
-
-				if curIndex == followFrom && curEntries != nil {
-					// Note that it's possible for Spanner to automatically retry transactions in some circumstances, when it does
-					// it'll call this function again.
-					// If the above condition holds, then we're in a retry situation and we must use the same data again rather
-					// than continue reading entries which will take us out of sync.
-				} else {
-					bs := uint64(batchSize)
-					if r := size - followFrom; r < bs {
-						bs = r
-					}
-					batch := make([][]byte, 0, bs)
-					for i := 0; i < int(bs); i++ {
-						idx, c, err := entryReader.Next()
-						if err != nil {
-							return fmt.Errorf("entryReader.next: %v", err)
-						}
-						if wantIdx := followFrom + uint64(i); idx != wantIdx {
-							// We're out of sync
-							return errOutOfSync
-						}
-						batch = append(batch, c)
-					}
-					curEntries = batch
-					curIndex = followFrom
-				}
-
-				// Store antispam entries.
-				//
-				// Note that we're writing the antispam entries outside of the transaction here. The reason is because we absolutely do not want
-				// the transaction to fail if there's already an entry for the same hash in the IDSeq table.
-				//
-				// It looks unusual, but is ok because:
-				//  - individual antispam entries fails because there's already an entry for that hash is perfectly ok
-				//  - we'll only continue on to update FollowCoord if no errors (other than AlreadyExists) occur while inserting entries
-				//  - similarly, if we manage to insert antispam entries here, but then fail to update FollowCoord, we'll end up
-				//    retrying over the same set of log entries, and then ignoring the AlreadyExists which will occur.
-				//
-				// Alternative approaches are:
-				//  - Use InsertOrUpdate, but that will keep updating the index associated with the ID hash, and we'd rather keep serving
-				//    the earliest index known for that entry.
-				//  - Perform reads for each of the hashes we're about to write, and use that to filter writes.
-				//    This would work, but would also incur an extra round-trip of data which isn't really necessary but would
-				//    slow the process down considerably and add extra load to Spanner for no benefit.
-				{
-					m := make([]*spanner.MutationGroup, 0, len(curEntries))
-					for i, e := range curEntries {
-						m = append(m, &spanner.MutationGroup{
-							Mutations: []*spanner.Mutation{spanner.Insert("IDSeq", []string{"id", "h", "idx"}, []interface{}{0, e, int64(curIndex + uint64(i))})},
-						})
-					}
-
-					i := d.dbPool.BatchWrite(ctx, m)
-					err := i.Do(func(r *spannerpb.BatchWriteResponse) error {
-						s := r.GetStatus()
-						if c := codes.Code(s.Code); c != codes.OK && c != codes.AlreadyExists {
-							return fmt.Errorf("failed to write antispam record: %v (%v)", s.GetMessage(), c)
-						}
-						return nil
-					})
-					if err != nil {
-						return err
-					}
-				}
-
-				numAdded := uint64(len(curEntries))
-				d.numWrites.Add(numAdded)
-
-				// Insertion of dupe entries was successful, so update our follow coordination row:
-				m := make([]*spanner.Mutation, 0)
-				m = append(m, spanner.Update("FollowCoord", []string{"id", "nextIdx"}, []interface{}{0, int64(followFrom + numAdded)}))
-
-				return txn.BufferWrite(m)
-			})
-			if err != nil {
-				if err != errOutOfSync {
-					klog.Errorf("Failed to commit antispam population tx: %v", err)
-				}
-				stop()
-				entryReader = nil
-				continue
-			}
-			curEntries = nil
-		}
-	}
 }
 
 // MigrationTarget creates a new GCP storage for the MigrationTarget lifecycle mode.
